requirements/local.txt
# Local development dependencies go here
-r test.txt
django-debug-toolbar==1.4.0
Sphinx==1.3.1
